all: fix int/double type mismatch between NED parameter and C++ usage

+++ src/inet/linklayer/csma/Csma.h
@@ -260,7 +260,7 @@ class INET_API Csma : public MacProtocolBase, public IMacProtocol
 
     /** @brief initial contention window size
      * Only used for linear and constant backoff method.*/
-    double initialCW;
+    int initialCW;
 
     /** @brief The power (in mW) to transmit with.*/
     double txPower;
